Update docs, show only updated apps
In the previous commit, all updated packages were shown in the sidebar,
instead of apps. This commit reverts that change, because apps are
usually more relevant to users.

It may make more sense to display recently updated packages on the page
that showcase packages, but that can be done later.