Prevent packages being overwritten by src architecture
With the new loading code introduced in 0da67c5, package were
overwritten by their "src" variant, since for each package there is
always a "src" architecture. Therefore, the web UI would only show the
"src" variant of a package and not the other architectures such as
"aarch64" or "i486".

This happened because before 0da67c5, all packages were added to a
list with a package for each architecture and later merged. After that
commit, each package was added to a dictionary with the package name as
a key. If multiple architectures for the same package were present, the
most recent would overwrite the old one. Since it seems like the
repository information always contains a package with the "src"
architecture, and it appears after all other architectures, it was the
only package architecture being registered.

Additionally, the package URL for the "src" architecture is now pulled
from the first repository, just like the "noarch" packages. Otherwise,
it would generate a lot of warnings and missing download links.