Try to parse changelog lines and format them on the web
Parse the lines of a changelog entry using the somewhat standardised
method of writing changelog messages. The parsed messages will be
displayed in a nicely formatted way.

If the parser cannot make any sense out of it, it will fall back to
displaying the changelog in a good, old <pre>-block.